# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Building Charges & Payments API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /persons:
    get:
      operationId: listPersons
      tags: [persons]
      summary: List persons
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
            enum: [owner, tenant]
        - name: nameAr
          in: query
          schema:
            type: string
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: includeArchived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
    post:
      operationId: createPerson
      tags: [persons]
      summary: Create a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /persons/{id}:
    get:
      operationId: getPerson
      tags: [persons]
      summary: Get a person by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Person found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          description: Not found
    patch:
      operationId: updatePerson
      tags: [persons]
      summary: Update a person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
    delete:
      operationId: archivePerson
      tags: [persons]
      summary: Archive a person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived

  # --- CHARGES ---
...
  /import/preview:
    post:
      operationId: previewImport
      tags: [import]
      summary: Parse and preview Excel import without saving
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportFileInput"
      responses:
        "200":
          description: Preview data with validation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreview"
  /import/commit:
    post:
      operationId: commitImport
      tags: [import]
      summary: Commit validated import data to database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCommitInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # --- EXPORT ---
...
components:
  schemas:
...
    Person:
      type: object
      required: [id, nameAr, role, unitId]
      properties:
        id:
          type: integer
        unitId:
          type: integer
        unitRef:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        buildingId:
          type: ["integer", "null"]
        buildingNameAr:
          type: ["string", "null"]
        nameAr:
          type: string
        role:
          type: string
          enum: [owner, tenant]
        phone:
          type: ["string", "null"]
        archived:
          type: boolean
        createdAt:
          type: string
    PersonInput:
      type: object
      required: [nameAr, role, unitId]
      properties:
        nameAr:
          type: string
        role:
          type: string
          enum: [owner, tenant]
        unitId:
          type: integer
        phone:
          type: string
    PersonUpdate:
      type: object
      properties:
        nameAr:
          type: string
        role:
          type: string
          enum: [owner, tenant]
        phone:
          type: string
        unitId:
          type: integer
...
    ImportFileInput:
      type: object
      properties:
        filename:
          type: string
    ImportRow:
      type: object
      properties:
        rowIndex:
          type: integer
        floor:
          type: ["integer", "null"]
        unitRef:
          type: string
        nameAr:
          type: string
        role:
          type: string
        category:
          type: ["string", "null"]
        tier:
          type: ["string", "null"]
        jan:
          type: ["number", "null"]
        feb:
          type: ["number", "null"]
        mar:
          type: ["number", "null"]
        apr:
          type: ["number", "null"]
        may:
          type: ["number", "null"]
        jun:
          type: ["number", "null"]
        jul:
          type: ["number", "null"]
        aug:
          type: ["number", "null"]
        sep:
          type: ["number", "null"]
        oct:
          type: ["number", "null"]
        nov:
          type: ["number", "null"]
        dec:
          type: ["number", "null"]
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    ImportPreview:
      type: object
      required: [rows, totalRows, validRows, errorRows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
        totalRows:
          type: integer
        validRows:
          type: integer
        errorRows:
          type: integer
        warnings:
          type: array
          items:
            type: string
    ImportCommitInput:
      type: object
      required: [rows, buildingId, year]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
        buildingId:
          type: integer
        year:
          type: integer
    ImportResult:
      type: object
      required: [unitsCreated, personsCreated, chargesCreated, errors]
      properties:
        unitsCreated:
          type: integer
        personsCreated:
          type: integer
        chargesCreated:
          type: integer
        errors:
          type: array
          items:
            type: string
...